updates to main App and account services

+++ capstone/tenmo-server/src/main/java/com/techelevator/tenmo/dao/JdbcTransferDao.java
@@ -75,7 +75,7 @@ public class JdbcTransferDao implements TransferDao {
         transfer.setTransferStatusId(rs.getInt("transfer_status_id"));
         transfer.setAccountFrom(rs.getInt("account_from"));
         transfer.setAccountTo(rs.getInt("account_to"));
-        transfer.setAmount(rs.getDouble("amount"));
+        transfer.setAmount(rs.getBigDecimal("amount"));
 
         return transfer;
     }
